chore(util): rename helpers dir to util
that ava thing: https://github.com/root-systems/dogstack/issues/75
